Add template keywords filter
